+++ fireants/registration/syn.py
@@ -77,7 +77,7 @@ class SyNRegistration(AbstractRegistration, DeformableMixin):
         affine (torch.Tensor): Initial affine transformation matrix
         smooth_warp_sigma (float): Smoothing sigma for warp field
     """
-    def __init__(self, scales: List[int], iterations: List[float], 
+    def __init__(self, scales: List[float], iterations: List[int], 
                 fixed_images: BatchedImages, moving_images: BatchedImages,
                 loss_type: str = "cc",
                 deformation_type: str = 'compositive',
